fix: Linter and code warnings (#18)
* :truck:: Move Grafana configuration to Kubernetes directory and add Vagrant README

The Grafana configuration file is moved to the Kubernetes directory to align with infrastructure changes, and a README file is added to the Vagrant folder to provide instructions and explanations about its contents.

Changes:
- Moved Grafana configuration file from "vm/grafana.yaml" to "ansible/files/k8s/grafana.yaml"
- Added a README file in the "vm" directory to explain the contents and purpose of the Vagrant folder
- Deleted the "vm/README.md" file which contained outdated information about Vagrant setup
- Deleted the "vm/vagrant/install-guest-additions.sh" script as it's no longer needed for the updated setup

* chore: Update CODE_OF_CONDUCT.md and SECURITY.md

Updated the CODE_OF_CONDUCT.md and SECURITY.md files to improve readability, clarity, and consistency. Changes include:

- Fixing formatting and alignment issues in the CODE_OF_CONDUCT.md file
- Updating URLs in the CODE_OF_CONDUCT.md and SECURITY.md files to use markdown hyperlink syntax for consistency and readability
- Adding line breaks for improved readability in the SECURITY.md file

* :art:: Update ansible-lint.yml and inventory YAML files

Made style adjustments to improve consistency and readability in the ansible-lint.yml file and YAML inventory files. Changes include:

- Excluded the "files" directory in ansible-lint.yml to align with project conventions
- Updated indentation and removed unnecessary whitespace in the inventory YAML files

* fix: Update Grafana resource limits and disable service account token automounting

This commit updates the resource limits for the Grafana container in the `grafana.yaml` file. The CPU limit is set to "1" and the CPU request is set to "0.5". Additionally, the service account token automounting is disabled by setting `automountServiceAccountToken` to false.