feat: Generate CA and TLS certificates using Ansible
This commit introduces changes to generate Certificate Authority (CA) and TLS certificates using Ansible.
It includes tasks to create private keys, certificate signing requests (CSRs), and sign certificates for various components:

- Admin user
- Kube-controller-manager
- Kube-proxy
- Kube-scheduler
- Kubernetes API Server
- API Server Kubelet Client
- ETCD Server
- Service Account

The certificates are generated successfully and can be found in the specified location.
A command `cert_verify.sh` is provided to verify the certificates.